pyproject.toml
[project]
name = "llm-rag-from-yt"
version = "0.1.0"
description = "LLM RAG system for YouTube audio content processing"
authors = [
    {name = "DataEngy", email = "data.engy@gmail.com"}
]
# readme = "docs/README.md"
requires-python = ">=3.9,<3.13"
dependencies = [
    "openai>=1.0.0",
    "python-dotenv",
    "pyyaml",
    "yt-dlp",
    "numpy<2",  # Pin to v1.x for torch compatibility
    "fastapi",
    "uvicorn",
    "loguru",
    "chromadb",
    "gradio",
    "pydantic",
    "typer",
    "rich",
    "bash",
    "plotly",
    "pandas",
    "python-telegram-bot>=20.0",
    "dagster>=1.5.0",
    "dagster-webserver>=1.5.0",
    "psutil",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-cov",
    "pytest-asyncio",
    "ruff",
    "pre-commit",
    "mypy",
    "httpx",  # for FastAPI testing
]
ml = [
    "faster-whisper",
    "sentence-transformers",
    "scikit-learn",
    "torch==2.2.2",  # Use compatible version for x86_64 macOS
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/llm_rag_yt"]

[project.scripts]
llm-rag-yt = "llm_rag_yt.cli.main:app"

[tool.ruff]
target-version = "py39"

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
ignore = [
    "E501",  # line too long
    "B008",  # do not perform function calls in argument defaults
    "C901",  # too complex
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]
"tests/*" = ["E501"]
"Pipeline_Demo.py" = ["ALL"]  # Exclude notebook conversion

[tool.ruff.format]

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra -q --strict-markers"
testpaths = ["tests"]
pythonpath = ["src"]
markers = [
    "asyncio: marks tests as requiring asyncio",
]

[tool.coverage.run]
source = ["src"]
omit = ["tests/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if self.debug:",
    "if settings.DEBUG",
    "raise AssertionError",
    "raise NotImplementedError",
    "if 0:",
    "if __name__ == .__main__.:",
]

[tool.mypy]
python_version = "3.9"
check_untyped_defs = true
disallow_any_generics = true
disallow_incomplete_defs = true
disallow_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
